[sil-verifier] Extract the flow sensitive verifier into its own file.
This has grown large enough and complex enough that it makes sense to go into
its own file.

This is a NFCI change. The only substantial changes is I added a small require
impl based on the one in SILVerifier and I eliminated the namespace
VerifyFlowSensitiveRulesDetails.
