+++ lib/SIL/Verifier/CMakeLists.txt
@@ -1,5 +1,6 @@
 target_sources(swiftSIL PRIVATE
   DebugInfoVerifier.cpp
+  FlowSensitiveVerifier.cpp
   LinearLifetimeChecker.cpp
   MemoryLifetimeVerifier.cpp
   GuaranteedPhiVerifier.cpp
